remove CORS middleware
CORS shouldn't be needed since the server is called directly by the ID Broker, not using XHR requests in a browser. If I am wrong and it is called in an XHR request, then the CORS Access-Control-Allow-Origin should be set to a specific value and not simply the origin value in the request header.